pyproject.toml
[project]
name = "project_generator"
authors = [
    {name = "Yu Wang", email = "thisisyuwang@gmail.com"},
]
version = "0.0.0-beta"
description = "A tool to quickly generate a project."
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "PyYAML",
]

[build-system]
requires = ["setuptools",]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
# All the following settings are optional:
where=['project_generator', '.']
include = ['*']  # ["*"] by default
exclude = ["tests*", 'config.yaml', '.gitignore', '__pyche__*']  # empty by default
namespaces = true  # true by default

[tool.setuptools.package-data]
templates = ["*.yaml", "*.json",]

[project.scripts]
project-generator = "project_generator.main:main"